Moved check_strategy to the verify_symbolic module
